Deduplicate symlink test helper
