+++ test/common/validation/test_val_input_file_or_dir_path.py
@@ -8,6 +8,7 @@ from pathlib import Path
 
 import pytest
 
+from scinoephile.common.testing import _create_symlink
 from scinoephile.common.validation import val_input_file_or_dir_path
 
 
@@ -47,21 +48,3 @@ def test_val_input_file_or_dir_path_resolves_symlink(tmp_path: Path):
     result = val_input_file_or_dir_path(symlink_path)
 
     assert result == file_path.resolve()
-
-
-def _create_symlink(
-    symlink_path: Path, target_path: Path, *, target_is_directory: bool = False
-):
-    """Create a symlink, skipping when Windows privileges do not allow it.
-
-    Arguments:
-        symlink_path: path at which to create the symlink
-        target_path: symlink target path
-        target_is_directory: whether the target is a directory
-    """
-    try:
-        symlink_path.symlink_to(target_path, target_is_directory=target_is_directory)
-    except OSError as exc:
-        if getattr(exc, "winerror", None) == 1314:
-            pytest.skip("Windows symlink privilege is not available")
-        raise
