+++ test/common/validation/test_val_input_path.py
@@ -9,6 +9,7 @@ from pathlib import Path
 import pytest
 
 from scinoephile.common.exceptions import NotAFileError
+from scinoephile.common.testing import _create_symlink
 from scinoephile.common.validation import val_input_path
 
 
@@ -157,21 +158,3 @@ def test_val_input_path_expands_vars(tmp_path: Path, monkeypatch: pytest.MonkeyP
     result = val_input_path("$TEST_FILE")
     assert result.exists()
     assert result.is_file()
-
-
-def _create_symlink(
-    symlink_path: Path, target_path: Path, *, target_is_directory: bool = False
-):
-    """Create a symlink, skipping when Windows privileges do not allow it.
-
-    Arguments:
-        symlink_path: path at which to create the symlink
-        target_path: symlink target path
-        target_is_directory: whether the target is a directory
-    """
-    try:
-        symlink_path.symlink_to(target_path, target_is_directory=target_is_directory)
-    except OSError as exc:
-        if getattr(exc, "winerror", None) == 1314:
-            pytest.skip("Windows symlink privilege is not available")
-        raise
